Updated the neural network

+++ NeuralNetwork/NeuralNetwork/NeuralNetwork/Optimizers.cpp
@@ -1,10 +1,10 @@
 #include "Optimizers.h"
 
-Optimizer::Optimizer(const int& learningRate_):
+Optimizer::Optimizer(const double& learningRate_):
 	learningRate{learningRate_}
 {}
 
-SGD::SGD(const int& learningRate_):
+SGD::SGD(const double& learningRate_):
 	Optimizer{learningRate_}
 {}
 
@@ -13,11 +13,11 @@ void SGD::update(MatrixXd& weights_, const MatrixXd& gradient_)
 	weights_ -= learningRate * gradient_;
 }
 
-SGDMomentum::SGDMomentum(const int& learningRate_, const double& gamma_):
+SGDMomentum::SGDMomentum(const double& learningRate_, const double& gamma_):
 	Optimizer{learningRate_}, gamma{gamma_}
 {}
 
-SGDMomentum::SGDMomentum(const int& learningRate_):
+SGDMomentum::SGDMomentum(const double& learningRate_):
 	SGDMomentum{learningRate_,0.9}
 {}
 
@@ -27,11 +27,11 @@ void SGDMomentum::update(MatrixXd& weights_, const MatrixXd& gradient_)
 	weights_ -= weightsVelocity;
 }
 
-RMSProp::RMSProp(const int& learningRate_, const double& rho_):
+RMSProp::RMSProp(const double& learningRate_, const double& rho_):
 	Optimizer{learningRate_}, rho{rho_}, epsilon{1e-8}
 {}
 
-RMSProp::RMSProp(const int& learningRate_) :
+RMSProp::RMSProp(const double& learningRate_) :
 	RMSProp{learningRate_ ,0.7}
 {}
 
@@ -43,11 +43,11 @@ void RMSProp::update(MatrixXd& weights_, const MatrixXd& gradient_)
 	weights_  -= gradient_.cwiseProduct(radicalSt);
 }
 
-Adam::Adam(const int& learningRate_, const double& beta1_, const double& beta2_):
+Adam::Adam(const double& learningRate_, const double& beta1_, const double& beta2_):
 	Optimizer{learningRate_}, beta1{beta1_}, beta2{beta2_}, epsilon{1e-8}
 {}
 
-Adam::Adam(const int& learningRate_) :
+Adam::Adam(const double& learningRate_) :
 	Adam{learningRate_,0.9,0.999}
 {}
 
@@ -62,7 +62,7 @@ void Adam::update(MatrixXd& weights_, const MatrixXd& gradient_)
 	weights_ -= learningRate * firstMomentum.cwiseProduct(secondMomentumInv);
 }
 
-AdaGrad::AdaGrad(const int& learningRate_):
+AdaGrad::AdaGrad(const double& learningRate_):
 	Optimizer{learningRate_}, epsilon{1e-8}
 {}
